Refactor(Kitchenpos): Add Infrastructure package
