+++ src/main/java/kitchenpos/menu/domain/MenuGroupDao.java
@@ -1,7 +1,5 @@
 package kitchenpos.menu.domain;
 
-import kitchenpos.menu.domain.MenuGroup;
-
 import java.util.List;
 import java.util.Optional;
