+++ src/main/java/kitchenpos/menu/domain/MenuProductDao.java
@@ -1,7 +1,5 @@
 package kitchenpos.menu.domain;
 
-import kitchenpos.menu.domain.MenuProduct;
-
 import java.util.List;
 import java.util.Optional;
