+++ src/main/java/kitchenpos/product/domain/ProductDao.java
@@ -1,7 +1,5 @@
 package kitchenpos.product.domain;
 
-import kitchenpos.product.domain.Product;
-
 import java.util.List;
 import java.util.Optional;
