+++ src/main/java/kitchenpos/table/application/TableGroupService.java
@@ -1,11 +1,11 @@
 package kitchenpos.table.application;
 
 import kitchenpos.order.domain.OrderDao;
-import kitchenpos.table.domain.OrderTableDao;
-import kitchenpos.table.domain.TableGroupDao;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.OrderTableDao;
 import kitchenpos.table.domain.TableGroup;
+import kitchenpos.table.domain.TableGroupDao;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.CollectionUtils;
