+++ src/main/java/kitchenpos/table/domain/OrderTableDao.java
@@ -1,7 +1,5 @@
 package kitchenpos.table.domain;
 
-import kitchenpos.table.domain.OrderTable;
-
 import java.util.List;
 import java.util.Optional;
